Iprove conversions and refactor code
