+++ src/earthkit/utils/module.py
@@ -1,0 +1,14 @@
+# (C) Copyright 2025 ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+#
+
+import sys
+
+
+def is_module_loaded(module_name: str) -> bool:
+    return module_name in sys.modules
